Revert character code from core foundation to UInt8
I changed it to UInt16 to account for certain characters in the
previous commit, but I am unsure characters that are more than 8 bits
are allowed in paths.  I should think about this further, however
perhaps something to consider while I am working on #5

+++ src/utils/darwin.jl
@@ -71,7 +71,7 @@ end
 # https://developer.apple.com/documentation/corefoundation/1542730-cfstringgetcharacteratindex
 function _cfstring_get_character_at_index(cfstr::Cstring, idx::T) where {T <: Integer}
     # UniChar CFStringGetCharacterAtIndex(CFStringRef theString, CFIndex idx);
-    return Char(ccall(:CFStringGetCharacterAtIndex, UInt16, (Cstring, UInt32), cfstr, idx))
+    return Char(ccall(:CFStringGetCharacterAtIndex, UInt8, (Cstring, UInt32), cfstr, idx))
 end
 
 # https://github.com/vovkasm/input-source-switcher/blob/c5bab3de716db5e3dae3703ed3b72f2bf1cd51d3/utils.cpp#L9-L18
